Version 2.4.0, cleaned code, raised CMake to 3.14

+++ src/dBTable.h
@@ -13,11 +13,11 @@ extern float dBTable[dBTableSize];
 inline float dBApprox(float x, float y) 
 {
   if(x < y) {
-    return dBTable[lrintf(x/y*dBTableScale)];
+    return dBTable[lrintf(x/y*static_cast<float>(dBTableScale))];
   } else if(x == 0.0f) {
     return 0;
   } else {
-    return dBTable[lrintf(y/x*dBTableScale)];
+    return dBTable[lrintf(y/x*static_cast<float>(dBTableScale))];
   }
 }
